requirements.txt
streamlit
PyPDF2
cohere
python-dotenv